Create new folder to contain create-new-feature.  Update name and command to wvus-dev-utilities. Update init question on command run.
