+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "dev-utils",
+  "name": "wvus-dev-utilities",
   "version": "1.0.0",
   "description": "\"git init\" on steroids",
   "main": "index.js",
@@ -23,10 +23,6 @@
     "touch": "^1.0.0"
   },
   "bin": {
-    "dev-utils": "./index.js"
-  },
-  "repository" : {
-    "type" : "git",
-    "url" : "https://github.com/reescott/dev-utilities.git"
+    "wvus-dev-utilities": "./index.js"
   }
 }
